big changes to handler and added a docker controller
