+++ hpotter-server/hpotter/plugins/handler.py
@@ -10,7 +10,6 @@ from OpenSSL import crypto
 
 from hpotter.env import logger
 from hpotter.plugins.generic import PipeThread
-from hpotter.plugins import ssh, telnet
 
 global set_cert
 set_cert = False
@@ -76,7 +75,7 @@ class Plugin:
     def remove_support_file_structure(self):
         try:
             for cmd in self.teardown:
-                logger.info("---%s is removing the %s directory", self.name, cmd)
+                logger.info("%s is removing the %s directory", self.name, cmd)
                 os.rmdir(cmd)
         except FileExistsError:
             pass
@@ -128,6 +127,7 @@ class Plugin:
         t = self.connect_to_hpotter()
 
         self.instances.append((c, t))
+        logger.info('added a %s container' % self.name)
 
     def remove_instance(self, client, network):
         if self.total_active > 0:
@@ -155,143 +155,6 @@ class Plugin:
                 logger.info("removing TLS cert and key")
         except:
             raise FileNotFoundError
-
-
-def check_docker():
-    try:
-        s = subprocess.check_output('docker ps', shell=True)
-    except subprocess.CalledProcessError:
-        print("Ensure Docker is running, and please try again.")
-        sys.exit()
-
-def start_plugins():
-    # ensure Docker is running
-    try:
-        s = subprocess.check_output('docker ps', shell=True)
-    except subprocess.CalledProcessError:
-        print("Ensure Docker is running, and please try again.")
-        sys.exit()
-
-    config = read_in_config()
-    start_services(parse_services(config[0]))
-    all_plugins = parse_plugins(config[1])
-    current_thread = None
-    current_container = None
-
-    for plugin in all_plugins:
-        if plugin is not None:
-            try:
-
-                check_certs(plugin.cert)
-                client = docker.from_env()
-
-                container = plugin.container
-                if platform.machine() == 'armv6l':
-                    container = plugin.alt_container
-
-                try:
-                    for cmd in plugin.setup['mkdir']:
-                        logger.info("%s created the %s directory",
-                                    plugin.name, cmd)
-                        os.mkdir(cmd)
-                except FileExistsError:
-                    pass
-                except OSError as error:
-                    logger.info(error)
-                    return
-
-                if (plugin.volumes):
-                    current_container = client.containers.run(container, \
-                        detach=plugin.detach, ports=plugin.makeports(), \
-                        environment=[plugin.environment])
-
-                else:
-                    current_container = client.containers.run(container, \
-                        detach=plugin.detach, ports=plugin.makeports(), \
-                        read_only=True)
-
-                logger.info('Created: %s', plugin.name)
-                network.connect(current_container)
-                logger.info('Connected %s to %s network', plugin.name, network.name)
-            except OSError as err:
-
-                logger.info(err)
-                if current_container:
-                    logger.info(current_container.logs())
-                    # rm_container()
-                return
-
-            di = lambda a: re.sub(b'([\x00-\x20]|[\x7f-xff])+', b' ', a)
-
-            current_thread = PipeThread((plugin.listen_address, \
-                plugin.listen_port), (plugin.ports['connect_address'], \
-                plugin.ports['connect_port']), plugin.table, \
-                plugin.capture_length, request_type=plugin.request_type, tls=plugin.tls)
-
-            current_thread.start()
-            p_dict = {
-                "plugin": plugin,
-                "container": current_container,
-                "thread": current_thread
-            }
-            Singletons.active_plugins[plugin.name] = p_dict
-        else:
-            logger.info(
-                "yaml configuration seems to be missing some important information")
-
-
-def stop_plugins():
-    ssh.stop_server()
-    telnet.stop_server()
-    remove_certs()
-
-    for name, item in Singletons.active_plugins.items():
-        try:
-            for cmd in item["plugin"].teardown['rmdir']:
-                logger.info("---%s is removing the %s directory", name, cmd)
-                os.rmdir(cmd)
-        except FileExistsError:
-            pass
-        except FileNotFoundError:
-            pass
-        except OSError as error:
-            logger.info(name + ": " + str(error))
-            return
-        if item["container"] is not None:
-            item["thread"].request_shutdown()
-        logger.info("--- removing %s container", item["plugin"].name)
-        network.disconnect(item["container"].name, True)
-        network.reload()
-
-        #avoid race conditions between singletons
-        lock = threading.Lock()
-        lock.acquire()
-
-        #remove network once all containers are disconnected
-        if not network.containers:
-            network.remove()
-            logger.info("--- network removed")
-            lock.release()
-        logger.info("--- %s container disconnected from %s", item["plugin"].name, network.name)
-        item["container"].stop()
-        logger.info("--- %s container removed", item["plugin"].name)
-
-
-
-def parse_services(data):
-    list = []
-    for s in data:
-        for k, v in s.items():
-            list.append(Service(k, v['address'], v['port']))
-    return list
-
-def start_services(service_config):
-    for service in service_config:
-        if service.name == 'ssh':
-            return ssh.start_server(service.address, service.port)
-        if service.name == 'telnet':
-            return telnet.start_server(service.address, service.port)
-
 
 def create_tls_cert_and_key(tmp_file):
     key = crypto.PKey()
