+++ hpotter-server/hpotter/plugins/telnet.py
@@ -66,13 +66,14 @@ class TelnetHandler(socketserver.BaseRequestHandler):
 class TelnetServer(socketserver.ThreadingMixIn, socketserver.TCPServer): pass
 
 def start_server(address='0.0.0.0', port=23):
+    global telnet_server
     telnet_handler = TelnetHandler
     telnet_server = TelnetServer((address, port), telnet_handler)
     threading.Thread(target=telnet_server.serve_forever).start()
     logger.info("Telnet server is up and running")
     return telnet_server
 
-def stop_server(s):
-    if s:
-        s.shutdown()
+def stop_server():
+    if telnet_server:
+        telnet_server.shutdown()
         logger.info("Telnet server was shutdown")
